feat: finished all handlers and storage for now

+++ pkg/upload/s3.go
@@ -2,29 +2,38 @@ package upload
 
 import (
 	"bytes"
+	"context"
 	"fmt"
+	"io"
 	"net/http"
+	"net/url"
 	"path"
 	"strings"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/session"
-	s3client "github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	awsconfig "github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/credentials"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
+	"github.com/aws/aws-sdk-go-v2/service/s3/types"
+	transport "github.com/aws/smithy-go/endpoints"
+	"github.com/gabriel-vasile/mimetype"
 	"github.com/kleister/kleister-api/pkg/config"
 )
 
 // S3Upload implements the Upload interface.
 type S3Upload struct {
-	endpoint string
-	path     string
-	access   string
-	secret   string
-	bucket   string
-	region   string
+	endpoint  string
+	path      string
+	access    string
+	secret    string
+	bucket    string
+	region    string
+	pathstyle bool
+	proxy     bool
 
-	client *s3client.S3
+	client  *s3.Client
+	presign *s3.PresignClient
 }
 
 // Info prepares some informational message about the handler.
@@ -35,19 +44,18 @@ func (u *S3Upload) Info() map[string]interface{} {
 	result["path"] = u.path
 	result["bucket"] = u.bucket
 	result["region"] = u.region
+	result["pathstyle"] = u.pathstyle
+	result["proxy"] = u.proxy
 
 	return result
 }
 
 // Prepare simply prepares the upload handler.
 func (u *S3Upload) Prepare() (Upload, error) {
-
-	cfg := aws.NewConfig().
-		WithRegion(u.region).
-		WithS3ForcePathStyle(true)
-
-	if u.endpoint != "" {
-		cfg = cfg.WithEndpoint(u.endpoint)
+	globalOpts := []func(*awsconfig.LoadOptions) error{
+		awsconfig.WithRegion(
+			u.region,
+		),
 	}
 
 	if u.access != "" && u.secret != "" {
@@ -63,22 +71,72 @@ func (u *S3Upload) Prepare() (Upload, error) {
 			return nil, fmt.Errorf("failed to parse secret key: %w", err)
 		}
 
-		cfg = cfg.WithCredentials(credentials.NewStaticCredentials(
-			access,
-			secret,
-			"",
-		))
+		globalOpts = append(
+			globalOpts,
+			awsconfig.WithCredentialsProvider(
+				credentials.NewStaticCredentialsProvider(
+					access,
+					secret,
+					"",
+				),
+			),
+		)
 	}
 
-	sess, err := session.NewSession()
+	cfg, err := awsconfig.LoadDefaultConfig(
+		context.Background(),
+		globalOpts...,
+	)
 
 	if err != nil {
-		return u, err
+		return nil, fmt.Errorf("failed to load configuration: %w", err)
 	}
 
-	u.client = s3client.New(
-		sess,
+	serviceOpts := []func(*s3.Options){
+		func(o *s3.Options) {
+			o.UsePathStyle = u.pathstyle
+		},
+	}
+
+	if u.endpoint != "" {
+		endpoint, err := url.Parse(u.endpoint)
+
+		if err != nil {
+			return nil, err
+		}
+
+		serviceOpts = append(
+			serviceOpts,
+			s3.WithEndpointResolverV2(
+				&CustomEndpointResolver{
+					Endpoint:  endpoint,
+					Bucket:    u.bucket,
+					PathStyle: u.pathstyle,
+				},
+			),
+		)
+
+		if endpoint.Scheme == "http" {
+			serviceOpts = append(
+				serviceOpts,
+				func(o *s3.Options) {
+					o.HTTPClient = &http.Client{
+						Transport: &http.Transport{
+							TLSClientConfig: nil,
+						},
+					}
+				},
+			)
+		}
+	}
+
+	u.client = s3.NewFromConfig(
 		cfg,
+		serviceOpts...,
+	)
+
+	u.presign = s3.NewPresignClient(
+		u.client,
 	)
 
 	return u, nil
@@ -90,16 +148,29 @@ func (u *S3Upload) Close() error {
 }
 
 // Upload stores an attachment within the defined S3 bucket.
-func (u *S3Upload) Upload(key, ctype string, content []byte) error {
-	params := &s3client.PutObjectInput{
-		ACL:         aws.String("public-read"),
+func (u *S3Upload) Upload(ctx context.Context, key string, content *bytes.Buffer) error {
+	reader := bytes.NewReader(
+		content.Bytes(),
+	)
+
+	mtype, err := mimetype.DetectReader(
+		reader,
+	)
+
+	if err != nil {
+		return err
+	}
+
+	params := &s3.PutObjectInput{
+		ACL:         types.ObjectCannedACLPublicRead,
 		Bucket:      aws.String(u.bucket),
 		Key:         aws.String(path.Join(u.path, key)),
-		ContentType: aws.String(ctype),
-		Body:        bytes.NewReader(content),
+		ContentType: aws.String(mtype.String()),
+		Body:        reader,
 	}
 
 	if _, err := u.client.PutObject(
+		ctx,
 		params,
 	); err != nil {
 		return err
@@ -109,16 +180,71 @@ func (u *S3Upload) Upload(key, ctype string, content []byte) error {
 }
 
 // Delete removes an attachment from the defined S3 bucket.
-func (u *S3Upload) Delete(key string) error {
-	params := &s3client.DeleteObjectInput{
-		Bucket: aws.String(u.bucket),
-		Key:    aws.String(path.Join(u.path, key)),
-	}
+func (u *S3Upload) Delete(ctx context.Context, key string, recursive bool) error {
+	if recursive {
+		var (
+			continuation *string
+		)
 
-	if _, err := u.client.DeleteObject(
-		params,
-	); err != nil {
-		return err
+		for {
+			objects, err := u.client.ListObjectsV2(
+				ctx, &s3.ListObjectsV2Input{
+					Bucket:            aws.String(u.bucket),
+					Prefix:            aws.String(path.Join(u.path, key)),
+					ContinuationToken: continuation,
+				},
+			)
+
+			if err != nil {
+				return fmt.Errorf("failed to list objects: %w", err)
+			}
+
+			if len(objects.Contents) == 0 {
+				break
+			}
+
+			deletable := make([]types.ObjectIdentifier, 0)
+
+			for _, object := range objects.Contents {
+				deletable = append(
+					deletable,
+					types.ObjectIdentifier{
+						Key: object.Key,
+					},
+				)
+			}
+
+			if _, err = u.client.DeleteObjects(
+				ctx,
+				&s3.DeleteObjectsInput{
+					Bucket: aws.String(u.bucket),
+					Delete: &types.Delete{
+						Objects: deletable,
+						Quiet:   aws.Bool(true),
+					},
+				},
+			); err != nil {
+				return fmt.Errorf("failed to delete objects: %w", err)
+			}
+
+			if !aws.ToBool(objects.IsTruncated) {
+				break
+			}
+
+			continuation = objects.NextContinuationToken
+		}
+	} else {
+		params := &s3.DeleteObjectInput{
+			Bucket: aws.String(u.bucket),
+			Key:    aws.String(path.Join(u.path, key)),
+		}
+
+		if _, err := u.client.DeleteObject(
+			ctx,
+			params,
+		); err != nil {
+			return err
+		}
 	}
 
 	return nil
@@ -127,40 +253,120 @@ func (u *S3Upload) Delete(key string) error {
 // Handler implements an HTTP handler for asset uploads.
 func (u *S3Upload) Handler(root string) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		req, _ := u.client.GetObjectRequest(&s3client.GetObjectInput{
-			Bucket: aws.String(u.bucket),
-			Key:    aws.String(path.Join(u.path, strings.TrimPrefix(r.URL.Path, root))),
-		})
-
-		url, err := req.Presign(5 * time.Minute)
-
-		if err != nil {
-			http.Error(
-				w,
-				http.StatusText(http.StatusNotFound),
-				http.StatusNotFound,
-			)
+		if u.proxy {
+			u.proxyHandler(root, w, r)
+		} else {
+			u.presignHandler(root, w, r)
 		}
-
-		http.Redirect(w, r, url, http.StatusTemporaryRedirect)
 	})
+}
+
+func (u *S3Upload) proxyHandler(root string, w http.ResponseWriter, r *http.Request) {
+	obj := strings.TrimPrefix(
+		path.Join(
+			u.path,
+			strings.TrimPrefix(
+				r.URL.Path,
+				root,
+			),
+		),
+		"/",
+	)
+
+	req, err := u.client.GetObject(
+		r.Context(),
+		&s3.GetObjectInput{
+			Bucket: aws.String(u.bucket),
+			Key:    aws.String(obj),
+		},
+	)
+
+	if err != nil {
+		http.Error(
+			w,
+			http.StatusText(http.StatusNotFound),
+			http.StatusNotFound,
+		)
+
+		return
+	}
+
+	defer req.Body.Close()
+
+	modified := time.Now()
+
+	if req.LastModified != nil {
+		modified = *req.LastModified
+	}
+
+	buffer, err := io.ReadAll(req.Body)
+
+	if err != nil {
+		http.Error(
+			w,
+			"Failed to read buffer",
+			http.StatusInternalServerError,
+		)
+
+		return
+	}
+
+	http.ServeContent(
+		w,
+		r,
+		obj,
+		modified,
+		bytes.NewReader(buffer),
+	)
+}
+
+func (u *S3Upload) presignHandler(root string, w http.ResponseWriter, r *http.Request) {
+	obj := strings.TrimPrefix(
+		path.Join(
+			u.path,
+			strings.TrimPrefix(
+				r.URL.Path,
+				root,
+			),
+		),
+		"/",
+	)
+
+	req, err := u.presign.PresignGetObject(
+		r.Context(),
+		&s3.GetObjectInput{
+			Bucket: aws.String(u.bucket),
+			Key:    aws.String(obj),
+		},
+		func(opts *s3.PresignOptions) {
+			opts.Expires = time.Duration(5 * time.Minute)
+		},
+	)
+
+	if err != nil {
+		http.Error(
+			w,
+			http.StatusText(http.StatusNotFound),
+			http.StatusNotFound,
+		)
+
+		return
+	}
+
+	http.Redirect(w, r, req.URL, http.StatusTemporaryRedirect)
 }
 
 // NewS3Upload initializes a new S3 handler.
 func NewS3Upload(cfg config.Upload) (Upload, error) {
-	path := "/"
-
-	if cfg.Path != "" {
-		path = cfg.Path
-	}
-
 	f := &S3Upload{
-		endpoint: cfg.Endpoint,
-		path:     path,
-		access:   cfg.Access,
-		secret:   cfg.Secret,
-		bucket:   cfg.Bucket,
-		region:   cfg.Region,
+		endpoint:  cfg.Endpoint,
+		pathstyle: cfg.Pathstyle,
+		path:      cfg.Path,
+		access:    cfg.Access,
+		secret:    cfg.Secret,
+		bucket:    cfg.Bucket,
+		region:    cfg.Region,
+		proxy:     cfg.Proxy,
 	}
 
 	return f.Prepare()
@@ -175,4 +381,26 @@ func MustS3Upload(cfg config.Upload) Upload {
 	}
 
 	return db
+}
+
+// CustomEndpointResolver is used for S3 compatible storage endpoints.
+type CustomEndpointResolver struct {
+	Endpoint  *url.URL
+	Bucket    string
+	PathStyle bool
+}
+
+// ResolveEndpoint resolves endpoints for a specific service and region
+func (r *CustomEndpointResolver) ResolveEndpoint(_ context.Context, _ s3.EndpointParameters) (transport.Endpoint, error) {
+	endpoint := r.Endpoint
+
+	if r.PathStyle {
+		endpoint = endpoint.JoinPath(
+			r.Bucket,
+		)
+	}
+
+	return transport.Endpoint{
+		URI: *endpoint,
+	}, nil
 }
